pyproject.toml
[project]
name = "oikia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "black>=25.1.0",
    "celery>=5.5.2",
    "fastapi>=0.115.12",
    "flake8>=7.2.0",
    "httpx>=0.28.1",
    "isort>=6.0.1",
    "pydantic-settings>=2.9.1",
    "pydantic[email]>=2.11.4",
    "pyjwt>=2.10.1",
    "pytest>=8.3.5",
    "python-json-logger>=3.3.0",
    "python-multipart>=0.0.20",
    "tortoise-orm>=0.25.0",
    "typer>=0.15.4",
    "uvicorn>=0.34.2",
    "valkey>=6.1.0",
]

[tool.black]
line-length = 180
exclude = '''
/(
    \.git
  | __pycache__
  | old
  | build
  | dist
  | venv
  | frontend
)/
'''
[tool.isort]
profile = "black"
